requirements.txt
openai-whisper
pyannote.audio
python-dotenv
numpy
torch
torchaudio
